+++ shell.core/src/main/java/org/metaborg/spoofax/shell/client/Repl.java
@@ -1,17 +1,12 @@
-package org.metaborg.spoofax.shell.client.console;
+package org.metaborg.spoofax.shell.client;
 
 import java.io.IOException;
 
-import org.fusesource.jansi.Ansi;
-import org.metaborg.spoofax.shell.client.IDisplay;
-import org.metaborg.spoofax.shell.client.IEditor;
 import org.metaborg.spoofax.shell.commands.CommandNotFoundException;
 import org.metaborg.spoofax.shell.commands.ICommandInvoker;
 import org.metaborg.spoofax.shell.commands.IReplCommand;
 
-import com.google.inject.Guice;
 import com.google.inject.Inject;
-import com.google.inject.Injector;
 import com.google.inject.Provider;
 
 /**
@@ -19,13 +14,7 @@ import com.google.inject.Provider;
  * via an {@link IEditor}, executes them with a {@link ICommandInvoker} and prints it to an
  * {@link IDisplay}.
  */
-public final class Repl {
-    private static Injector injector;
-
-    static {
-        injector = Guice.createInjector(new ReplModule());
-    }
-
+public class Repl {
     private ICommandInvoker invoker;
     private IEditor editor;
     private IDisplay display;
@@ -54,9 +43,6 @@ public final class Repl {
      *             when an IO error occurs.
      */
     public void run() throws IOException {
-        display.displayResult(Ansi.ansi().a("Welcome to the ").bold().a("Spoofax").reset()
-            .a(" REPL").toString());
-
         String input;
         running = true;
         while ((input = editor.getInput()) != null && running) {
@@ -96,15 +82,5 @@ public final class Repl {
         public void execute(String... args) {
             replProvider.get().running = false;
         }
-    }
-
-    /**
-     * @param args
-     *            Unused.
-     * @throws IOException
-     *             when an IO error occurs.
-     */
-    public static void main(String[] args) throws IOException {
-        injector.getInstance(Repl.class).run();
     }
 }
